Added react router dom and css for various components
