+++ src/App.jsx
@@ -1,22 +1,17 @@
-import Navbar from './Components/Navbar/Navbar';
-import Footer from './Components/Footer/Footer';
-import About from './Components/About/About';
-import Appointment from './Components/Appointment/Appointment';
-import ContactUs from './Components/ContactUs/ContactUs';
-import Blog from './Components/Blog/Blog';
-import Services from './Components/Services/Services';
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import './App.css'
+import Home from "./pages/home";
+import Blogs from "./pages/blogs";
 function App() {
   return (
     <div className='App'>
-      <Navbar/>
-      <About/>
-      <Appointment />
-      <Services />
-      <Blog />
-      <ContactUs />
-      <Footer/>
-    </div> 
+      <Router>
+        <Routes>
+          <Route path="/" exact element={<Home />} />
+          <Route path="/blogs" exact element={<Blogs />} />
+        </Routes>
+      </Router>
+    </div>
   )
 }
