+++ src/Components/ContactUs/style.css
@@ -1,8 +1,3 @@
-.address-container {
-    color: white;
-    border-radius: 10px;
-}
-
 .business-contact {
     font-weight: bold;
 }
