+++ src/Components/Services/ServiceCard/style.css
@@ -5,7 +5,7 @@ a,p,h1,h2,h3,h4,ul {
 
 .section-services {
 	/* padding-top: 110px; */
-	padding-bottom: 80px;
+	/* padding-bottom: 80px; */
 	font-family: "Poppins", sans-serif;
 	/* background-color: #211f24; */
 	color: #fff;
@@ -54,14 +54,20 @@ a,p,h1,h2,h3,h4,ul {
 }
 
 .section-services .single-service {
-	margin-top: 40px;
+	/* margin-top: 20px; */
 	background-color: #24252a;
 	box-shadow: 0 0 10px 0 rgba(0,0,0,.1);
+	border-radius: 10px;
+	display: flex;
+    align-items: flex-start;
+    flex-direction: column;
 }
 
 .section-services .single-service .part-1 {
-	padding: 40px 40px 25px;
+	padding: 20px;
 	border-bottom: 2px solid #1d1e23;
+	flex-grow: 2;
+	width: 100%;
 }
 
 /*change services icon css in this class*/
@@ -79,7 +85,8 @@ a,p,h1,h2,h3,h4,ul {
 }
 
 .section-services .single-service .part-2 {
-	padding: 30px 40px 40px;
+	padding: 20px;
+	width: 100%;
 }
 
 .section-services .single-service .part-2 .description {
