+++ src/pages/blogs.jsx
@@ -1,0 +1,15 @@
+import Blog from '../Components/Blog/Blog'
+import Footer from '../Components/Footer/Footer'
+import BlogNavbar from '../Components/Navbar/BlogNavbar'
+
+function Blogs() {
+    return (
+        <>
+            <BlogNavbar />
+            <Blog />
+            <Footer />
+        </>
+    )
+}
+
+export default Blogs
